Added copious casts to avoid possible integer overflows in the Microsoft Windows 64-bit build, where sizeof(long) < sizeof(size_t).

+++ coders/art.c
@@ -105,12 +105,12 @@ static Image *ReadARTImage(const ImageInfo *image_info,ExceptionInfo *exception)
   dummy=ReadBlobLSBShort(image);
   height=ReadBlobLSBShort(image);
 
-  ldblk=(long) ((width+7) / 8);
+  ldblk=(long) (((size_t) width+7) / 8);
   Padding=(unsigned char) ((-ldblk) & 0x01);
 
   image->columns=width;
   image->rows=height;
-  if(GetBlobSize(image)!=(magick_off_t) (8+((long)ldblk+Padding)*image->rows))
+  if(GetBlobSize(image)!=(magick_off_t) (8+((size_t)ldblk+Padding)*image->rows))
     ThrowReaderException(CorruptImageError,ImproperImageHeader,image);
   if (CheckImagePixelLimits(image, exception) != MagickPass)
     ThrowReaderException(ResourceLimitError,ImagePixelLimitExceeded,image);
